Refactor : App 구조 리팩토링 및 역할별 파일 분리
- App.tsx에서 라우팅 관련 코드를 AppRoutes.tsx로 분리
- FullPageLayout 컴포넌트를 layouts 디렉토리로 이동
- 페이지와 컴포넌트 간 역할 명확화
- 코드 가독성과 유지보수성을 위한 폴더 구조 개선
