+++ src/components/admin/AdminComponent.tsx
@@ -1,27 +1,13 @@
-import { useNavigate } from "react-router-dom";
 import { touristInfo, touristSpots, weatherData } from "../../data/adminData";
 import {WeatherCard} from "./cards/weatherCard";
 import {SpotCard} from "./cards/spotCard";
+import AdminHeader from "./AdminHeader";
 
-export default function Admin() {
-    const navigate = useNavigate();
+export default function AdminComponent() {
 
     return (
         <div className="bg-gray-100 h-auto flex flex-col w-full overflow-y-hidden">
-            <div className="bg-white px-6 py-4 flex items-center relative">
-                <button
-                    className="absolute left-6 bg-white shadow-md px-4 py-2 text-indigo-500 font-semibold hover:bg-indigo-500 hover:text-white transition"
-                    onClick={() => {
-                        alert("보안을 위해 자동 로그아웃됩니다.");
-                        navigate("/");
-                    }}
-                >
-                    ← 돌아가기
-                </button>
-                <h1 className="text-2xl text-center font-bold text-gray-600 w-full">
-                    STARS 관리자 통합 화면
-                </h1>
-            </div>
+            <AdminHeader />
 
             <div className="flex flex-1 overflow-hidden p-4">
                 <div
