+++ src/components/admin/AdminLogin.tsx
@@ -46,7 +46,7 @@ export default function AdminLogin() {
                                 id="username"
                                 name="username"
                                 type="text"
-                                placeholder={"Admin ID"}
+                                placeholder={"AdminComponent ID"}
                                 required
                                 className="bg-white appearance-none rounded relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm"
                                 value={username}
