+++ src/components/stars/dashboard/POICardList.tsx
