+++ src/components/stars/dashboard/WeatherCard.tsx
