--- src/components/ui/Card.tsx
@@ -1,11 +1,0 @@
-export function Card({ className = '', children }: { className?: string; children?: React.ReactNode }) {
-    return (
-        <div className={`bg-white rounded-xl shadow p-4 ${className}`}>
-            {children}
-        </div>
-    );
-}
-
-export function CardContent({ className = '', children }: { className?: string; children?: React.ReactNode }) {
-    return <div className={`p-2 ${className}`}>{children}</div>;
-}
