+++ src/pages/admin/AdminLoginPage.tsx
@@ -1,0 +1,9 @@
+import AdminLogin from "../../components/admin/AdminLogin";
+
+export default function AdminLoginPage() {
+    return (
+        <div>
+            <AdminLogin />
+        </div>
+    );
+}
