+++ src/pages/admin/AdminPage.tsx
@@ -1,0 +1,9 @@
+import AdminComponent from "../../components/admin/AdminComponent";
+
+export default function AdminPage() {
+    return (
+        <div>
+            <AdminComponent />
+        </div>
+    );
+}
