+++ src/pages/stars/DashboardPage.tsx
@@ -1,0 +1,9 @@
+import DashboardComponent from "../../components/stars/DashboardComponent";
+
+export default function HomePage() {
+    return (
+        <div>
+            <DashboardComponent />
+        </div>
+    );
+}
