+++ src/pages/stars/MapSectionPage.tsx
@@ -1,0 +1,9 @@
+import MapSectionComponent from "../../components/stars/MapSectionComponent";
+
+export default function MapSectionPage() {
+    return (
+        <div>
+            <MapSectionComponent />
+        </div>
+    );
+}
